#70 Implemented deleteTransferTransaction, etc.
- Implemented deleteTransferTransaction from dao to usecase
- Added a notATransferTransaction error

- Allow enable/disable of top/bottom divider for TransferTransactionListTile
- Renamed BottomSheetContent to TxnBottomSheet and updated its props
- Added TransferTxnBottomSheet
- Minor change to error UI for TransferTransactionListTile

- Minor changes to DebugTransactionsScreen following the prop change of TxnBottomSheet
