+++ lib/presentation/debug/debug_transactions/debug_transactions_screen.dart
@@ -9,10 +9,9 @@ import 'package:pecunia/features/accounts/domain/entities/account.dart';
 import 'package:pecunia/features/accounts/usecases/get_all_accounts.dart';
 import 'package:pecunia/features/transactions/domain/entities/transaction.dart';
 import 'package:pecunia/features/transactions/usecases/create_transaction.dart';
-import 'package:pecunia/features/transactions/usecases/delete_transaction.dart';
 import 'package:pecunia/features/transactions/usecases/get_all_transactions.dart';
+import 'package:pecunia/presentation/debug/debug_accounts/view_account/txn_bottom_sheet_widget.dart';
 import 'package:pecunia/presentation/debug/debug_forms/create_txn_form_widget.dart';
-import 'package:pecunia/presentation/debug/debug_forms/edit_txn_form_widget.dart';
 import 'package:pecunia/presentation/dialogs/pecunia_dialogs.dart';
 
 class DebugTransactionsScreen extends ConsumerWidget {
@@ -131,6 +130,15 @@ class ViewAllTransactions extends ConsumerWidget {
             child: Text('No transactions yet!'),
           );
         }
+
+        final account = accountsList.whenData((list) {
+          if (list.isEmpty) {
+            throw Exception('No accounts found, but transactions exist!');
+          }
+
+          return list.firstWhere((element) => element.id == transactions.first.accountId);
+        }).asData!;
+
         return Padding(
           padding: const EdgeInsets.symmetric(horizontal: 8),
           child: ListView.builder(
@@ -170,8 +178,7 @@ class ViewAllTransactions extends ConsumerWidget {
                           ),
                     trailing: BuildTxnAmountText(txn),
                     onTap: () {
-                      showTransactionBottomSheet(
-                          context, txn, accountsList.whenData((value) => value).asData?.value ?? []);
+                      showTransactionBottomSheet(context, txn, account.value);
                     },
                   ),
                   Divider(color: Colors.grey.withOpacity(0.1)),
@@ -199,350 +206,4 @@ class ViewAllTransactions extends ConsumerWidget {
     final account = accounts.firstWhere((element) => element.id == accountId);
     return account.name;
   }
-}
-
-class BuildTxnAmountText extends ConsumerWidget {
-  const BuildTxnAmountText(this.txn, {super.key});
-
-  final Transaction txn;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final isCredit = txn.fundDetails.transactionType == TransactionType.credit;
-    final isMultiCurrency = txn.fundDetails.isMultiCurrency;
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.end,
-      children: [
-        Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            if (isCredit)
-              Text(
-                '+${txn.fundDetails.transactionAmount}',
-                style: TextStyle(
-                  color: Colors.green[300],
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            if (!isCredit)
-              Text(
-                '-${txn.fundDetails.transactionAmount}',
-                style: TextStyle(
-                  color: Colors.red[300],
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            const SizedBox(width: 10),
-            Text(
-              txn.fundDetails.transactionCurrency.code,
-              style: TextStyle(
-                color: Colors.grey[600],
-                fontSize: 14,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ],
-        ),
-        if (isMultiCurrency) const SizedBox(height: 4),
-        if (isMultiCurrency)
-          Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              if (isCredit)
-                Text(
-                  '+${txn.fundDetails.exchangedAmount}',
-                  style: TextStyle(
-                    color: Colors.green[300]!.withOpacity(0.3),
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              if (!isCredit)
-                Text(
-                  '-${txn.fundDetails.exchangedAmount}',
-                  style: TextStyle(
-                    color: Colors.red[300]!.withOpacity(0.3),
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              const SizedBox(width: 10),
-              Text(
-                txn.fundDetails.exchangedCurrency.code,
-                style: TextStyle(
-                  color: Colors.grey[600]!.withOpacity(0.3),
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ],
-          ),
-      ],
-    );
-  }
-}
-
-class BottomSheetContent extends ConsumerWidget {
-  const BottomSheetContent(this.txn, this.accountsList, this.account, {super.key});
-
-  final Transaction txn;
-  final List<Account> accountsList;
-  final Account? account;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final chosenAccount = account ?? accountsList.firstWhere((element) => element.id == txn.accountId);
-
-    return Container(
-      padding: const EdgeInsets.only(top: 14, bottom: 14),
-      child: Column(
-        children: [
-          ListTile(
-            title: Text(
-              txn.name,
-              style: const TextStyle(fontWeight: FontWeight.bold),
-            ),
-            subtitle: txn.transactionDescription.value == null
-                ? null
-                : Text(
-                    txn.transactionDescription.value!,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-            trailing: BuildTxnAmountText(txn),
-          ),
-          const Divider(),
-          Container(
-            alignment: Alignment.centerLeft,
-            width: double.infinity,
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 4),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                RichText(
-                  text: TextSpan(
-                    text: 'txn_id: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.orange[300],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(text: txn.id, style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'acc_id: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.orange[300],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(text: txn.accountId, style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'creator_id: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.orange[300],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(text: txn.creatorUid, style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'txn_type: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.orange[300],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.transactionType.typeAsString,
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'txn_date: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.orange[300],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.transactionDate.toString(), style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'txn_amount: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.deepPurple[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.transactionAmount.toString(),
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'exchange_rate: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.deepPurple[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.exchangeRate.toString(),
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'base_currency: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.blue[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.baseCurrency.code, style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'base_amount: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.blue[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.baseAmount.toString(),
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'target_currency: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.pink[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.targetCurrency?.code ?? 'null',
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    text: 'target_amount: ',
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          fontWeight: FontWeight.bold,
-                          color: Colors.pink[200],
-                        ),
-                    children: <TextSpan>[
-                      TextSpan(
-                          text: txn.fundDetails.targetAmount.toString(),
-                          style: DefaultTextStyle.of(context).style),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-          const Divider(),
-          ListTile(
-            title: const Text('Edit'),
-            leading: const Icon(Icons.edit),
-            onTap: () {
-              context.pop();
-              showEditTransactionBottomSheet(context, txn, chosenAccount);
-            },
-          ),
-          ListTile(
-            title: Text('Delete', style: TextStyle(color: Colors.red[300])),
-            leading: Icon(Icons.delete, color: Colors.red[300]),
-            onTap: () {
-              ref.read(pecuniaDialogsProvider).showConfirmationDialog(
-                    title: 'Delete transaction?',
-                    message: "This isn't a reversible action, think twice.",
-                    onConfirm: () {
-                      ref.read(deleteTransactionProvider.notifier).deleteTransaction(txn);
-                    },
-                    context: context,
-                  );
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-void showTransactionBottomSheet(BuildContext context, Transaction txn, List<Account> accountsList) {
-  showModalBottomSheet<void>(
-      isScrollControlled: true,
-      context: context,
-      showDragHandle: true,
-      builder: (context) {
-        return SizedBox(
-          height: 550,
-          child: SingleChildScrollView(
-            physics: const BouncingScrollPhysics(parent: NeverScrollableScrollPhysics()),
-            child: BottomSheetContent(txn, accountsList, null),
-          ),
-        );
-      });
-}
-
-void showEditTransactionBottomSheet(BuildContext context, Transaction txn, Account account) {
-  showModalBottomSheet<void>(
-      isScrollControlled: true,
-      context: context,
-      showDragHandle: true,
-      builder: (context) {
-        return SizedBox(
-          height: 700,
-          child: SingleChildScrollView(
-            physics: const BouncingScrollPhysics(parent: AlwaysScrollableScrollPhysics()),
-            child: Column(
-              children: [
-                EditTxnForm(txn: txn, account: account),
-              ],
-            ),
-          ),
-        );
-      });
 }
